can send messages to the bar now, yippee!

+++ src/ipc.c
@@ -1,17 +1,16 @@
 #include "ipc.h"
+#include "../utils/log.h"
 
 #include <stddef.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <assert.h>
 #include <errno.h>
-
-#include "../utils/log.h"
-
 #include <string.h>
+
 #include <sys/socket.h>
-#include <sys/un.h>
 #include <unistd.h>
+#include <fcntl.h>
 
 bool client_setup(struct bar_ipc *bar_ipc) {
 	struct sockaddr_un *sock = bar_ipc->socket;
@@ -91,7 +90,6 @@ bool server_setup(struct bar_ipc *bar_ipc) {
 
 	char *sock_path = "/tmp/scoopybar-socket";
 	strncpy(sock->sun_path, sock_path, sizeof(sock->sun_path) - 1);
-    sock->sun_family = AF_UNIX;
 	if (setenv("SCOOPYBARSOCK", sock_path, 1) == -1) {
 		log_err(__FILE__, __LINE__, "Failed to set socket.");
 		return false;
@@ -115,10 +113,18 @@ bool server_setup(struct bar_ipc *bar_ipc) {
 bool IPC_socket_init(struct bar_ipc *bar_ipc, enum sock_type type) {
 	assert( bar_ipc != NULL );
 	bar_ipc->socket_fd = socket(AF_UNIX, SOCK_STREAM, 0);
+    bar_ipc->socket->sun_family = AF_UNIX;
 	if (bar_ipc->socket_fd == -1) {
 		log_err(__FILE__, __LINE__, "Failed to create socket fd.");
 		return false;
 	}
+    if (fcntl(bar_ipc->socket_fd, F_SETFD, FD_CLOEXEC) == -1) {
+        log_err(__FILE__, __LINE__, "Failed to set FD_CLOEXEC.");
+        return false;
+    }
+    if (fcntl(bar_ipc->socket_fd, F_SETFL, O_NONBLOCK) == -1) {
+        log_err(__FILE__, __LINE__, "Failed to set nonblocking on fd.");
+    }
 
 	if (type == SERVER) {
 		return server_setup(bar_ipc);
@@ -132,26 +138,37 @@ void IPC_socket_destroy(struct bar_ipc *bar_ipc, enum sock_type type) {
 
     if (close(bar_ipc->socket_fd) == -1) log_err(__FILE__, __LINE__, "Failed to close socket fd.");
     if (bar_ipc->socket != NULL) {
-        unlink(bar_ipc->socket->sun_path);
+        if (type == SERVER) unlink(bar_ipc->socket->sun_path);
         free(bar_ipc->socket);
     }
     free(bar_ipc);
 
-    if (type == SERVER) {
+    if (type == SERVER) { // might not need anymore
         unsetenv("SCOOPYBARSOCK");
     }
 }
 
 bool bar_receive_msg(struct bar_ipc *server) {
     int accept_fd = accept(server->socket_fd, NULL, NULL);
+    if (accept_fd == -1) {
+        if (errno == EAGAIN || errno == EWOULDBLOCK) { // No connection present.
+            errno = 0;
+            return false;
+        }
+        log_err(__FILE__, __LINE__, "Failed to accept.");
+        return false;
+    }
 	size_t b_read = read(accept_fd, server->msg, sizeof(server->msg));
 	if (b_read == -1) {
 		log_err(__FILE__, __LINE__, "Error reading from socket.");
+        close(accept_fd);
 		return false;
 	}
 	if (b_read == 0) {
+        close(accept_fd);
 		return false; // Nothing sent.
 	}
+    close(accept_fd);
 	return true;
 }
